musicapi/requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
Django==4.0.5
django-filter==21.1
djangorestframework==3.13.1
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.1.1
pytz==2022.1
PyYAML==6.0
requests==2.27.1
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
wincertstore==0.2
